// src/types/v2/File/Pb/PbWorkSheet.ts
import { BaseWorksheet } from '@/types/v2/File/Common/BaseWorksheet';

export interface PbWorkSheet extends BaseWorksheet {
    generateur: string;
    marque: string;
    modele: string;
    puissance: string;
    conduiteMateriau: string;
    conduiteDiametre: string;
    longueurTotal: string;
    longeurProjection: string;
    nbCoude90: string;
    nbCoude45: string;
    reductionSection: boolean;
    etat: boolean;
    demontable: boolean;
    distanceSecurite: string;
    conduitType: string;
    conduitMateriauConstitutif: string;
    plaqueSignaletique: boolean;
    classeTemperature: string;
    classePression: string;
    resistanceCondansat: string;
    resistanceCorrosion: boolean;
    resistanceFeu: string;
    distanceSecuriteCombustible: string;
    presenceTrappe: boolean;
    hauteurTotal: string;
    hauteurLocauxChauffe: string;
    hauteurLocauxNonChauffe: string;
    hauteurExterieur: string;
    devoiement: boolean;
    distanceDevoiement: string;
    conduitIsole: boolean;
    sectionConduitLargeur: string;
    sectionConduitLongeur: string;
    sectionConduitDiametre: string;
    deboucheSup40: boolean;
    obstacleInf8: boolean;
    deboucheAccessible: boolean;
    typeDebouche: string;
    toiture: string;
    pieceLogement: string;
    pieceLongueur: string;
    pieceLargeur: string;
    pieceHauteur: string;
    pieceSurface: string;
    accesPorteLargeur: string;
    accesPorteHauteur: string;
    natureMur: string;
    natureSol: string;
    naturePlafond: string;
    ameneeAir: boolean;
    priseElec: string;
    niveauHabitation: string;
    typeChantier: string;
    escalier: boolean;
    escalierLargeur: string;
    zoneInstallation: string;
    creation: boolean;
}